+++ src/__tests__/restLink.ts
@@ -374,12 +374,13 @@ describe('Complex responses need nested __typename insertions', () => {
       key: string,
       __typename: string,
       patcher: RestLink.FunctionalTypePatcher,
+      context: RestLink.TypePatcherContext,
     ) => {
       const value = data[key];
       if (value == null) {
         return {};
       }
-      const result = { [key]: patcher(value, __typename, patcher) };
+      const result = { [key]: patcher(value, __typename, patcher, context) };
       return result;
     };
     const typePatcher: RestLink.TypePatcherTable = {
@@ -387,6 +388,7 @@ describe('Complex responses need nested __typename insertions', () => {
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -394,33 +396,42 @@ describe('Complex responses need nested __typename insertions', () => {
 
         return {
           ...obj,
-          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper),
+          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper, context),
           ...patchIfExists(
             obj,
             'simpleDoubleNesting',
             'SimpleDoubleNesting',
             patchDeeper,
+            context,
           ),
-          ...patchIfExists(obj, 'nestedArrays', 'NestedArrays', patchDeeper),
+          ...patchIfExists(
+            obj,
+            'nestedArrays',
+            'NestedArrays',
+            patchDeeper,
+            context,
+          ),
         };
       },
       Inner1: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
         }
         return {
           ...obj,
-          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper),
+          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper, context),
         };
       },
       SimpleDoubleNesting: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -428,13 +439,14 @@ describe('Complex responses need nested __typename insertions', () => {
 
         return {
           ...obj,
-          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper),
+          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper, context),
         };
       },
       NestedArrays: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -447,12 +459,14 @@ describe('Complex responses need nested __typename insertions', () => {
             'singlyArray',
             'SinglyNestedArrayEntry',
             patchDeeper,
+            context,
           ),
           ...patchIfExists(
             obj,
             'doublyNestedArray',
             'DoublyNestedArrayEntry',
             patchDeeper,
+            context,
           ),
         };
       },
@@ -653,12 +667,13 @@ describe('Complex responses need nested __typename insertions', () => {
       key: string,
       __typename: string,
       patcher: RestLink.FunctionalTypePatcher,
+      context: RestLink.TypePatcherContext,
     ) => {
       const value = data[key];
       if (value == null) {
         return {};
       }
-      const result = { [key]: patcher(value, __typename, patcher) };
+      const result = { [key]: patcher(value, __typename, patcher, context) };
       return result;
     };
     const typePatcher: RestLink.TypePatcherTable = {
@@ -666,6 +681,7 @@ describe('Complex responses need nested __typename insertions', () => {
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -673,33 +689,42 @@ describe('Complex responses need nested __typename insertions', () => {
 
         return {
           ...obj,
-          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper),
+          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper, context),
           ...patchIfExists(
             obj,
             'simpleDoubleNesting',
             'SimpleDoubleNesting',
             patchDeeper,
+            context,
           ),
-          ...patchIfExists(obj, 'nestedArrays', 'NestedArrays', patchDeeper),
+          ...patchIfExists(
+            obj,
+            'nestedArrays',
+            'NestedArrays',
+            patchDeeper,
+            context,
+          ),
         };
       },
       Inner1: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
         }
         return {
           ...obj,
-          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper),
+          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper, context),
         };
       },
       SimpleDoubleNesting: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -707,13 +732,14 @@ describe('Complex responses need nested __typename insertions', () => {
 
         return {
           ...obj,
-          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper),
+          ...patchIfExists(obj, 'inner1', 'Inner1', patchDeeper, context),
         };
       },
       NestedArrays: (
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
@@ -726,12 +752,14 @@ describe('Complex responses need nested __typename insertions', () => {
             'singlyArray',
             'SinglyNestedArrayEntry',
             patchDeeper,
+            context,
           ),
           ...patchIfExists(
             obj,
             'doublyNestedArray',
             'DoublyNestedArrayEntry',
             patchDeeper,
+            context,
           ),
         };
       },
@@ -739,13 +767,14 @@ describe('Complex responses need nested __typename insertions', () => {
         obj: any,
         outerType: string,
         patchDeeper: RestLink.FunctionalTypePatcher,
+        context: RestLink.TypePatcherContext,
       ) => {
         if (obj == null) {
           return obj;
         }
         return {
           ...obj,
-          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper),
+          ...patchIfExists(obj, 'reused', 'Reused', patchDeeper, context),
         };
       },
     };
@@ -3889,6 +3918,7 @@ describe('Apollo client integration', () => {
             data: any,
             outerType: string,
             patchDeeper: RestLink.FunctionalTypePatcher,
+            context: RestLink.TypePatcherContext,
           ): any => {
             // Let's make unfairCriticism a Required Field
             if (data.unfairCriticism == null) {